% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_repo.R
\name{download.repo}
\alias{download.repo}
\title{Download a GitHup Repo}
\usage{
download.repo(
  repo,
  githuburl = "https://github.com/RVerse-Tutorials/",
  clean = TRUE
)
}
\arguments{
\item{repo}{The name of the repo in quotes}

\item{clean}{If TRUE (default), remove .gitignore, .gitattributes, *.RProj}

\item{girhuburl}{The url of the GitHub account in quotes}
}
\description{
This downloads a repo and unzips into the working directory
}
